Modify Dockerfile for use multisage build
The multi stage is beter for the final image size, and much more.

Removed some minor caveats, for better usage in containers. The port
must be passed as a env on future for better container usage, but the
certs can be used as volumes now.

There is no need to add certs onimage, and set the exec bit on
dockerfile.